backend-django/requirements.app.txt
Django==5.1.2
whitenoise==6.8.2
python-dotenv==1.0.1
PyYAML==6.0.2
plotly==5.24.1
pandas==2.2.3
sqlparse==0.5.1
asgiref==3.8.1

